requirements.txt
PyYAML==3.11
chromalog==1.0.2
Jinja2==2.7.3
six==1.9.0
voluptuous==0.8.7
pyzmq==14.5.0
